refactor: Separate test page components
