+++ src/pages/test/TestCreate.tsx
@@ -1,0 +1,11 @@
+import MainTemplate from '@/components/MainTemplate';
+
+interface Props {
+  fmNo: number;
+}
+
+export default function TestCreate({ fmNo }: Props) {
+  return (
+    <MainTemplate title="Select Training Session">TestCreate</MainTemplate>
+  );
+}
